test: migrate e2e test to component test (refs SFKUI-6500)
